% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatSD.R
\name{MatSD}
\alias{MatSD}
\title{Vectorised Standard Deviation}
\usage{
MatSD(x, dim = 1L)
}
\arguments{
\item{x}{A Matrix}

\item{dim}{dimention: eather 1L for rowwsie or 2L for colwise}
}
\value{
rowwsie/colwise variance
}
\description{
Vectorised Standard Deviation
}
\examples{
set.seed(123)
M <- matrix(sample(25), ncol = 5)
MatSD(M, 1L)
}

