% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatVar.R
\name{MatVar}
\alias{MatVar}
\title{Vectorised Variance}
\usage{
MatVar(x, dim = 1L)
}
\arguments{
\item{x}{A Matrix}

\item{dim}{dimention: either 1L for rowwsie or 2L for colwise}
}
\value{
rowwsie/colwise variance
}
\description{
Vectorised Variance
}
\examples{
set.seed(123)
M <- matrix(sample(25), ncol = 5)
MatVar(M, 1L)
}

