% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowCor.R
\name{RowCor}
\alias{RowCor}
\title{Vectorised rowwsie Pearson Correlation}
\usage{
RowCor(A, B)
}
\arguments{
\item{A}{a Matrix}

\item{B}{a matrix}
}
\value{
rowwsie Correlation
}
\description{
Vectorised rowwsie Pearson Correlation
}
\examples{
set.seed(123)
A <- matrix(sample(25), ncol = 5)
B <- matrix(sample(25), ncol = 5)
RowCor(A, B)
}

