% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowCov.R
\name{RowCov}
\alias{RowCov}
\title{Vectorised rowwsie Covariance}
\usage{
RowCov(A, B)
}
\arguments{
\item{A}{a Matrix}

\item{B}{a Matrix}
}
\value{
rowwsie Covariance
}
\description{
Vectorised rowwsie Covariance
}
\examples{
set.seed(123)
A <- matrix(sample(25), ncol = 5)
B <- matrix(sample(25), ncol = 5)
RowCov(A, B)
}

