% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowMaxs.R
\name{RowMaxs}
\alias{RowMaxs}
\title{Vectorised Row Maxs}
\usage{
RowMaxs(x, na.rm = FALSE)
}
\arguments{
\item{x}{A Matrix}

\item{na.rm}{If you want to remove NAs TRUE/FALSE}
}
\value{
rowwsie maxs
}
\description{
Vectorised Row Maxs
}
\examples{
set.seed(123)
M <- matrix(sample(c(NA, 1:20), 20, replace = TRUE), ncol = 5L)
RowMaxs(M, na.rm = TRUE)
}

