% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowMins.R
\name{RowMins}
\alias{RowMins}
\title{Vectorised Row Mins}
\usage{
RowMins(x, na.rm = FALSE)
}
\arguments{
\item{x}{A Matrix}

\item{na.rm}{If you want to remove NAs TRUE/FALSE}
}
\value{
rowwsie mins
}
\description{
Vectorised Row Mins
}
\examples{
set.seed(123)
M <- matrix(sample(c(NA, 1:20), 20, replace = TRUE), ncol = 5L)
RowMins(M, na.rm = TRUE)
}

