% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowSD.R
\name{RowSD}
\alias{RowSD}
\title{Vectorised rowwsie Standard Deviation}
\usage{
RowSD(x)
}
\arguments{
\item{x}{A Matrix}
}
\value{
rowwsie Standard Deviation
}
\description{
Vectorised rowwsie Standard Deviation
}
\examples{
set.seed(123)
M <- matrix(sample(25), ncol = 5)
RowSD(M)
}

