% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowVar.R
\name{RowVar}
\alias{RowVar}
\title{Vectorised rowwsie Variance}
\usage{
RowVar(x)
}
\arguments{
\item{x}{A Matrix}
}
\value{
rowwsie variance
}
\description{
Vectorised rowwsie Variance
}
\examples{
set.seed(123)
M <- matrix(sample(25), ncol = 5)
RowVar(M)
}

